+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## [UNRELEASED]
+## [0.13.7]
 ### Added
  - Implemented cash rounding for currency formatting.
  - Currency options now accepts `cash: boolean` to enable cash rounding.
@@ -99,6 +99,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 #### Fixed/Changed
 - Converted `CLDR` to an interface
 
+[0.13.7]: https://github.com/phensley/cldr-engine/compare/v0.13.6...v0.13.7
 [0.13.6]: https://github.com/phensley/cldr-engine/compare/v0.13.5...v0.13.6
 [0.13.5]: https://github.com/phensley/cldr-engine/compare/v0.13.4...v0.13.5
 [0.13.4]: https://github.com/phensley/cldr-engine/compare/v0.12.1...v0.13.4